cpp-proio: improved CMake library config stuff
* Renamed libproioProto.so to libproio.pb.so
* Added set(proio_LIBRARIES ...) to proioConfig.cmake
* Library dependencies specified such that if the user wants to use the
packaged data models, they must explicity add the proio.pb target as a
link dependency (in addition to the proio target).  Alternatively, the
user can just use the proio_LIBRARIES variable for explicit paths to
both libproio.so and libproio.pb.so.  However, in the latter case, the
user then needs to find and link the protobuf library.